refactor: move Nasdaq provider to provider/nasdaq sub-package
